Move BitSized to model
