+++ keelhaul/src/generate.rs
@@ -281,55 +281,13 @@ impl TestConfig {
     }
 }
 
-/// `BitSized` types have a knowable size
-pub(crate) trait BitSized<T>: fmt::Debug + Copy {
-    fn bit_count() -> u32;
-    fn all_ones() -> T;
-    fn can_represent<U: cmp::PartialOrd<T>>(val: U) -> bool {
-        val <= Self::all_ones()
-    }
-}
-
-impl BitSized<u8> for u8 {
-    fn bit_count() -> u32 {
-        8
-    }
-    fn all_ones() -> u8 {
-        u8::MAX
-    }
-}
-impl BitSized<u16> for u16 {
-    fn bit_count() -> u32 {
-        16
-    }
-    fn all_ones() -> u16 {
-        u16::MAX
-    }
-}
-impl BitSized<u32> for u32 {
-    fn bit_count() -> u32 {
-        32
-    }
-    fn all_ones() -> u32 {
-        u32::MAX
-    }
-}
-impl BitSized<u64> for u64 {
-    fn bit_count() -> u32 {
-        64
-    }
-    fn all_ones() -> u64 {
-        u64::MAX
-    }
-}
-
 /// Generates a "bitwise and" operation for given value and mask
 ///
 /// # Examples
 ///
 /// `val = 0xb0`, `mask = u8::MAX` -> `0xb0u8`
 /// `val = 0xb0`, `mask = 1` ->  `(0xb0u8 & 0b1u8)`
-fn gen_bitand<T: BitSized<T> + fmt::LowerHex + fmt::Binary + cmp::PartialOrd>(
+fn gen_bitand<T: model::BitSized<T> + fmt::LowerHex + fmt::Binary + cmp::PartialOrd>(
     value: T,
     mask: T,
 ) -> TokenStream {
@@ -349,7 +307,10 @@ fn gen_bitand<T: BitSized<T> + fmt::LowerHex + fmt::Binary + cmp::PartialOrd>(
 }
 
 /// Get a literal hexadecimal representation of `val`, e.g., "0xdeadbeef"
-fn u_to_hexlit<T: fmt::LowerHex + cmp::PartialOrd + BitSized<T>>(val: T, bits: u32) -> String {
+fn u_to_hexlit<T: fmt::LowerHex + cmp::PartialOrd + model::BitSized<T>>(
+    val: T,
+    bits: u32,
+) -> String {
     assert!(
         T::can_represent(val),
         "value `{val:?}` cannot be represented using `{}`",
@@ -365,7 +326,7 @@ fn u_to_hexlit<T: fmt::LowerHex + cmp::PartialOrd + BitSized<T>>(val: T, bits: u
 }
 
 /// Get a literal binary representation of `val`, e.g., "0b10101010"
-fn u_to_binlit<T: fmt::Binary + cmp::PartialOrd + BitSized<T>>(val: T, bits: u32) -> String {
+fn u_to_binlit<T: fmt::Binary + cmp::PartialOrd + model::BitSized<T>>(val: T, bits: u32) -> String {
     assert!(
         T::can_represent(val),
         "value `{val:?}` cannot be represented using `{}`",
